requirements.txt
numpy>=1.22
pandas>=1.5
matplotlib>=3.6
scikit-learn>=1.2
torch>=2.0
tqdm>=4.64
scipy>=1.9
pyserial>=3.5
